+++ CHANGELOG.md
@@ -4,9 +4,10 @@ Version 1.13 (not yet released)
 * Find contact by email or create with email and first/last name
 * Find or create contact by email and first/last name
 * Validate checksum
-* Add 'details' parameter to CreateActivity action 
-* Add action to retrieve the currently active/associated contact 
-* Add 'set preferred communication method' action 
+* Add 'details' parameter to CreateActivity action
+* Add action to retrieve the currently active/associated contact
+* Add 'set preferred communication method' action
+* Add 'Create relationship (with relationship type parameter)' action
 
 Version 1.12
 ------------