+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version 1.13 (not yet released)
 * Add 'Subscribe to mailing list' action
 * Add 'Confirm mailing list subscription' action
 * Changed Upload Custom File Field and Add Attachment action so it also accepts a URL for the attachment.
+* Add "Calculate value (binary arithmetic operation)" action
 
 Version 1.12
 ------------